Implement authorization on tenant and user operations
Defined SUPERUSER and USER role for service sysadmin and tenant users
respectively. Restricted tenant operations to SUPERUSER. Similarly
restricted user operations to SUPERUSER except for operations like
changing password on current user, which anyone can do.  Also extended
session so that it has tenantID as well as an effectiveTenantId. The
latter is a placeholder to permit sysadmin pose as a user for a particular
tenant.